@@ -6,7 +6,8 @@ package com.jdragon.tljrobot.client.config;
  */
 public class FinalConfig {
     public final static String SECRET = "NwobS3vT3wShr53wI2fvmVbOtV6I5dq9b2n6Zat2DZR";
-    public final static String VERSION = "v2.24";
+    public final static String VERSION = "v2.26";
     public final static String ENCODING = "utf8";
-
+    public final static String TLJ_ADDR = "https://tlj.wiki";
+    public final static String XL_ADDR = "http://xlboy.cn:8081";
 }
