@@ -19,23 +19,23 @@ import com.jdragon.tljrobot.robot.newTyping.tools.Regex;
 import com.jdragon.tljrobot.robot.typing.Tools.RegexText;
 import com.jdragon.tljrobot.tljutils.DateUtil;
 import com.jdragon.tljrobot.tljutils.HttpUtil;
-import org.junit.Test;
 
+import java.sql.Date;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
 /**
  * Create by Jdragon on 2020.01.21
+ * 该监听抓取与赛文有关
+ * 包括：
+ *  1、联赛，群日赛的成绩收取
+ *  2、联赛，日赛当日与历史成绩的图片发送
+ *  3、上传联赛赛文
  */
 public class RobotGroupClient extends IcqListener {
-    @Test
-    public void test(){
-        String[] Com = new RegexText().CarryCom("常用单字前五百\n" +
-                "微报老来怎\n" +
-                "-----第20638段-乱序--c.fw.1.5");
-        System.out.println(Com);
-    }
     @EventHandler
     public void carryGroupMessage(EventGroupMessage eventGroupMessage){
         String message = eventGroupMessage.getMessage();
@@ -50,9 +50,7 @@ public class RobotGroupClient extends IcqListener {
             JSONObject.parseObject(HttpUtil.doPostObjectAndParams(HttpAddr.UPLOAD_GROUP_ARTICLE_CACHE,postParams,article));
         }
         if (qq==robot.xiaochaiQ) {
-            System.out.println("捕获赛文");
             if ("999".equals(Com[2])&&GroupCache.typeGroupMap.containsKey(groupId)) {
-                System.out.println("捕获赛文");
                 JSONObject articleJson = new JSONObject();
                 articleJson.put("title", Com[0]);
                 articleJson.put("content",Com[1]);
@@ -68,29 +66,16 @@ public class RobotGroupClient extends IcqListener {
             }
         }else if("#联赛".equals(message)){
             //{"code":200,"message":"获取成功","result":{"id":0,"holdDate":"2020-01-22","articleId":1,"author":"321321","article":{"id":1,"title":"1","content":"1"}}}
-            JSONObject retJson = JSONObject.parseObject(HttpUtil.doPost(HttpAddr.GET_UNION_MATCH, DateUtil.now().toString()));
-            String retMessage = retJson.getString("message");
-            if("获取成功".equals(retMessage)) {
-                System.out.println(retJson);
-                JSONObject resultJson = retJson.getJSONObject("result");
-                JSONObject articleJson = resultJson.getJSONObject("article");
-                String author = resultJson.getString("author");
-                String title = articleJson.getString("title");
-                String content = articleJson.getString("content");
-                String respond = title+" 投稿自QQ号："+author+"\n"+content+"\n-----第9999段-共"+content.length()+"字";
-                eventGroupMessage.respond(respond);
-            }else{
-                eventGroupMessage.respond(retMessage);
-            }
+            eventGroupMessage.respond(getUnionArticle(DateUtil.now()));
         }else if("#联赛成绩".equals(message)){
-            eventGroupMessage.respond(resultUnionJson(false));
+            eventGroupMessage.respond(resultUnionJson(DateUtil.now(),false));
         }else if("#联赛首打".equals(message)){
-            eventGroupMessage.respond(resultUnionJson(true));
-        }else if ("#长流成绩".equals(message)){
+            eventGroupMessage.respond(resultUnionJson(DateUtil.now(),true));
+        }else if ("#长生成绩".equals(message)){
             //{"id":173,"userId":1,"typeDate":"2020-02-04","speed":123.0,"keySpeed":123.0,"keyLength":123.0,
             // "number":123,"deleteText":123,"deleteNum":123,"mistake":123,"repeatNum":123,"keyAccuracy":123.0,
             // "keyMethod":123.0,"wordRate":123.0,"time":123.0,"articleId":123,"paragraph":0,"userName":"谭宇"}
-            eventGroupMessage.respond(resultTljJson());
+            eventGroupMessage.respond(resultTljJson(DateUtil.now()));
         }else if(Regex.getParagraph(message)==999||Regex.getParagraph(message)==9999){
             if(!GroupCache.typeGroupMap.containsKey(groupId)) {
                 return;
@@ -107,6 +92,14 @@ public class RobotGroupClient extends IcqListener {
         }
         String[] params = message.split("\\s");
         if(params.length==2){
+            Date date;
+            try {
+                date = new Date(new SimpleDateFormat("yyyy-MM-dd").parse(RegexText.AddZero(params[1])).getTime());
+            } catch (ParseException e) {
+                eventGroupMessage.respond("日期格式错误:"+params[1]);
+                return;
+            }
+            System.out.println("判断"+date.toString());
             if("#历史赛文".equals(params[0])){
                 JSONObject jsonObject = JSON.parseObject(HttpUtil.doPost(HttpAddr.GET_GROUP_MATCH_ADDR, String.valueOf(groupId),params[1]));
                 String retMessage = jsonObject.getString("message");
@@ -119,6 +112,16 @@ public class RobotGroupClient extends IcqListener {
                 }else{
                     eventGroupMessage.respond(retMessage);
                 }
+            }else if("#联赛赛文".equals(params[0])){
+                eventGroupMessage.respond(getUnionArticle(date));
+            }else if("#联赛成绩".equals(params[0])){
+                eventGroupMessage.respond(resultUnionJson(date,false));
+            }else if("#联赛首打".equals(params[0])){
+                eventGroupMessage.respond(resultUnionJson(date,true));
+            }else if("#长生成绩".equals(params[0])){
+                eventGroupMessage.respond(resultTljJson(date));
+            }else if("#长生赛文".equals(params[0])){
+                eventGroupMessage.respond(getTljArticle(date));
             }
         }else if(params.length==3){
             if(groupId==robot.matchGroupNum&&"#投稿".equals(params[0])){
@@ -135,9 +138,40 @@ public class RobotGroupClient extends IcqListener {
             }
         }
     }
-    public static String resultTljJson(){
+    public static String getTljArticle(Date date){
+        JSONObject retJson = JSON.parseObject(HttpUtil.doGet(HttpAddr.MATCH_GET_TODAY, date.toString()));
+        String retMessage = retJson.getString("message");
+        if("获取成功".equals(retMessage)) {
+            JSONObject resultJson = retJson.getJSONObject("result");
+            JSONObject articleJson = resultJson.getJSONObject("article");
+            String author = resultJson.getString("author");
+            String title = articleJson.getString("title");
+            String content = articleJson.getString("content");
+            String respond = title+":"+author+"\n"+content+"\n-----第1段-共"+content.length()+"字";
+            return respond;
+        }else{
+            return retMessage;
+        }
+    }
+    public static String getUnionArticle(Date date){
+        JSONObject retJson = JSONObject.parseObject(HttpUtil.doPost(HttpAddr.GET_UNION_MATCH, date.toString()));
+        String retMessage = retJson.getString("message");
+        if("获取成功".equals(retMessage)) {
+            System.out.println(retJson);
+            JSONObject resultJson = retJson.getJSONObject("result");
+            JSONObject articleJson = resultJson.getJSONObject("article");
+            String author = resultJson.getString("author");
+            String title = articleJson.getString("title");
+            String content = articleJson.getString("content");
+            String respond = title+" 投稿自QQ号："+author+"\n"+content+"\n-----第9999段-共"+content.length()+"字";
+            return respond;
+        }else{
+            return retMessage;
+        }
+    }
+    public static String resultTljJson(Date date){
         String resoleResult;
-        String title = DateUtil.now().toString();
+        String title = date.toString();
         String[] head = new String[]{"序号","名字","成绩","击键","码长","退格","回改","错字","选重","键准","键法","打词","设备"};
         JSONObject retJson = JSONObject.parseObject(HttpUtil.doPost(HttpAddr.GET_TLJ_MATCH_RANK_ALL, title));
         String retMessage = retJson.getString("message");
@@ -159,12 +193,12 @@ public class RobotGroupClient extends IcqListener {
         }
         return resoleResult;
     }
-    public static String resultUnionJson(boolean isFirst){
+    public static String resultUnionJson(Date date,boolean isFirst){
         String resoleResult;
-        JSONObject articleJson = JSONObject.parseObject(HttpUtil.doPost(HttpAddr.GET_UNION_MATCH, DateUtil.now().toString()));
+        JSONObject articleJson = JSONObject.parseObject(HttpUtil.doPost(HttpAddr.GET_UNION_MATCH, date.toString()));
         String retMessage = articleJson.getString("message");
         if("获取成功".equals(retMessage)){
-            String title = articleJson.getJSONObject("result").getJSONObject("article").getString("title");
+            String title = date.toString()+":"+articleJson.getJSONObject("result").getJSONObject("article").getString("title");
             String[] head = new String[]{"序号","名字","来自","成绩","击键","码长","退格","回改","选重","错字","键准","重打"};
             /**
              * {
