@@ -4,8 +4,8 @@ package com.jdragon.tljrobot.robot.newTyping.config;
  * Create by Jdragon on 2020.01.21
  */
 public class HttpAddr {
-    private final static String SERVER_ADDR = "http://localhost:8081";
-//    private final static String SERVER_ADDR = "https://tlj.wiki";
+//    private final static String SERVER_ADDR = "http://localhost:8081";
+    private final static String SERVER_ADDR = "https://tlj.wiki";
 
     private final static String ROBOT_HOME = "/robot";
 
@@ -27,11 +27,15 @@ public class HttpAddr {
 
     public final static String GET_UNION_MATCH = SERVER_ADDR + ROBOT_HOME + UNION_ADDR + "/getUnionMatch";
 
+
+
     public final static String GET_UNION_FIRST_MATCH_RANK = SERVER_ADDR + ROBOT_HOME + UNION_ADDR + "/getUnionFirstAchRank";
 
     public final static String GET_UNION_MATCH_RANK = SERVER_ADDR + ROBOT_HOME + UNION_ADDR + "/getUnionAchRank";
 
     public final static String GET_TLJ_MATCH_RANK_ALL = SERVER_ADDR + ROBOT_HOME + TLJ_MATCH_ADDR + "/getTljMatchAchByDate";
+
+    public final static String MATCH_GET_TODAY = SERVER_ADDR + ROBOT_HOME + TLJ_MATCH_ADDR + "/getTljMatch";
 
     public final static String UPLOAD_GROUP_MATCH_ADDR = SERVER_ADDR + ROBOT_HOME +GROUP_MATCH_ADDR + "/uploadGroupMatch";
 
