@@ -2,39 +2,12 @@ package com.jdragon.tljrobot.robot.typing.ConDatabase;
 
 import com.jdragon.tljrobot.robot.typing.Tools.initGroupList;
 
-import java.sql.*;
-import java.util.Calendar;
-import java.util.GregorianCalendar;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 
 public class InConn {
-    public static void AddGroupSaiwen(Long groupid,String []Com){
-        try{
-            String sql = "SELECT * FROM groupsaiwen where groupid=? and saiwendate=?";
-            Date date = Conn.getdate();
-            Connection con = Conn.getConnection();
-            PreparedStatement ptmt = Conn.getPtmt(con,sql);
-            ptmt.setLong(1,groupid);
-            ptmt.setDate(2,date);
-            ResultSet rs = ptmt.executeQuery();
-            if(rs.next()){
-                con.close();
-                return;
-            }
-            else{
-                sql = "insert into groupsaiwen value(?,?,?,?)";
-                ptmt = Conn.getPtmt(con,sql);
-                ptmt.setLong(1,groupid);
-                ptmt.setDate(2,date);
-                ptmt.setString(3,Com[1]);
-                ptmt.setString(4,Com[0]);
-                ptmt.execute();
-                con.close();
-                System.out.println("添加赛文:"+groupid+"日期:"+ Conn.getdate());
-            }
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-    }
     public static void addMaxComMath(Long id,Long groupid,double[] Com){
         try{
             String sql = "select * from groupsaiwenmax where id=? and groupid=? and date=?";
@@ -144,240 +117,6 @@ public class InConn {
             e.printStackTrace();
         }
     }
-    public static String AddAllGroupSaiwenDate(String date,String title,String saiwen,long id){
-        String sql = "select * from allgroupsaiwen where saiwendate=?";
-        try{
-            Connection con = Conn.getConnection();
-            PreparedStatement ptst = Conn.getPtmt(con,sql);
-            ptst.setDate(1,Date.valueOf(date));
-            ResultSet rs = ptst.executeQuery();
-            if(rs.next()){
-                return "该日期已有赛文";
-            }else{
-                sql = "insert into allgroupsaiwen values(?,?,?,?)";
-                ptst = Conn.getPtmt(con,sql);
-                ptst.setDate(1,Date.valueOf(date));
-                ptst.setString(2,title);
-                ptst.setString(3,saiwen);
-                ptst.setLong(4,id);
-                ptst.execute();
-                con.close();
-                return "发布成功";
-            }
-        }catch (Exception e){
-            e.printStackTrace();
-            return "未知原因发布失败";
-        }
-    }
-    public static String AddAllGroupSaiwenWait(String title,String saiwen,long id){
-        String sql = "insert into allgroupsaiwenwait(title,saiwen,author) values(?,?,?)";
-        try{
-            Connection con = Conn.getConnection();
-            PreparedStatement ptst = Conn.getPtmt(con,sql);
-            ptst.setString(1,title);
-            ptst.setString(2,saiwen);
-            ptst.setLong(3,id);
-            ptst.execute();
-            con.close();
-            return "等待审核";
-        } catch (SQLException e) {
-            e.printStackTrace();
-            return "未知原因发布失败";
-        }
-    }
-    public static String WaitToUpload(Object[] QQandId,boolean agree){
-        int id = Integer.valueOf(QQandId[0].toString());
-        String sql = "select * from allgroupsaiwenwait where id="+id;
-        String str = "发布拒绝";
-        try{
-            Connection con = Conn.getConnection();
-            ResultSet rs = Conn.getStmtSet(con,sql);
-            if(rs.next()&&agree){
-                str = AddAllGroupSaiwen(rs.getString("title"),rs.getString("saiwen"),rs.getLong("author"));
-            }
-            sql = "delete from allgroupsaiwenwait where id="+id;
-            QQandId[1] = rs.getLong("author");
-            QQandId[2] = rs.getString("title");
-            Conn.execute(con,sql);
-            con.close();
-            return str;
-        } catch (SQLException e) {
-            e.printStackTrace();
-            return str;
-        }
-    }
-    public static String AddAllGroupSaiwen(String title,String saiwen,long id){
-        String sql = "select * from allgroupsaiwen ORDER BY saiwendate DESC";
-        try{
-            Connection con = Conn.getConnection();
-            ResultSet rs = Conn.getStmtSet(con,sql);
-            Date date ;
-            if(rs.next()){
-                date = rs.getDate("saiwendate");
-                if(Conn.getdate().getTime()>date.getTime())
-                    date = Conn.getdate();
-                else {
-                    Calendar calendar = new GregorianCalendar();
-                    calendar.setTime(date);
-                    calendar.add(calendar.DATE,1);
-                    java.util.Date date1 =  calendar.getTime();
-                    date = new Date(date1.getTime());
-                }
-            }
-            else date = Conn.getdate();
-            sql = "insert into allgroupsaiwen values(?,?,?,?)";
-            PreparedStatement ptst = Conn.getPtmt(con,sql);
-            ptst.setDate(1,date);
-            ptst.setString(2,title);
-            ptst.setString(3,saiwen);
-            ptst.setLong(4,id);
-            ptst.execute();
-            con.close();
-            return "审核通过赛文日期："+ date;
-//            }
-        }catch (Exception e){
-            e.printStackTrace();
-            return "未知原因发布失败";
-        }
-    }
-    public static String DeleteAllGroupSaiwenByDate(String date){
-        String sql = "delete from allgroupsaiwen where saiwendate=?";
-        try{
-            Connection con = Conn.getConnection();
-            PreparedStatement ptmt = Conn.getPtmt(con,sql);
-            ptmt.setDate(1,Date.valueOf(date));
-            int i = ptmt.executeUpdate();
-            if(i==0)return "删除失败";
-             else return "删除成功";
-        } catch (SQLException e) {
-            e.printStackTrace();
-            return "删除失败"+e.getMessage();
-        }
-    }
-    public static String ChangeAllGroupSaiwen(String date,String title,String saiwen,long id,boolean admin){
-        String sql = "select * from allgroupsaiwen where saiwendate=?";
-        String message = "";
-        try{
-            Connection con = Conn.getConnection();
-            PreparedStatement ptst = Conn.getPtmt(con,sql);
-            ptst.setDate(1,Date.valueOf(date));
-            ResultSet rs = ptst.executeQuery();
-            if(rs.next()){
-                if(rs.getLong("author")==id || admin){
-                    sql = "update allgroupsaiwen set title=?,saiwen=? where saiwendate=?";
-                    ptst = Conn.getPtmt(con,sql);
-                    ptst.setString(1,title);
-                    ptst.setString(2,saiwen);
-                    ptst.setDate(3,Date.valueOf(date));
-                    ptst.execute();
-                    message =  "修改成功";
-                }else {
-                    message =  "不是你发布的赛文，你无法修改";
-                }
-            }else{
-                message =  "该天无赛文，无法修改";
-            }
-            con.close();
-        }catch (Exception e){
-            e.printStackTrace();
-            message =  "未知原因发布失败";
-        }
-        return message;
-    }
-    public static String AddRobotSaiwen(String type,String title,String saiwen,long id,String nickname) {
-        String sql = "select * from robotsaiwen where type=? ORDER BY aid DESC";
-        try {
-            Connection con = Conn.getConnection();
-            PreparedStatement ptmt =  Conn.getPtmt(con,sql);
-            ptmt.setString(1,type);
-            ResultSet rs = ptmt.executeQuery();
-            int aid = 0;
-            if(rs.next())
-                aid = rs.getInt("aid");
-            sql = "insert into robotsaiwen values(?,?,?,?,?,?)";
-            ptmt = Conn.getPtmt(con,sql);
-            ptmt.setInt(1,++aid);
-            ptmt.setString(2,title);
-            ptmt.setString(3,saiwen);
-            ptmt.setString(4,nickname);
-            ptmt.setLong(5,id);
-            ptmt.setString(6,type);
-            ptmt.execute();
-            con.close();
-            return "发布成功文章编号：" + type+" "+aid;
-//            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            return "未知原因发布失败";
-        }
-    }
-    public static void addMaxAllGroupComMath(Long id,Long groupid,double[] Com,String name){
-        try{
-            String sql = "select * from allgroupsaiwenchengji where id=? and saiwendate=?";
-            String username = initGroupList.GroupMemberCardMap.get(groupid).get(id);
-            if(username.equals("")||username==null)username = name;
-            Connection con = Conn.getConnection();
-            Date date = Conn.getdate();
-            System.out.println(date);
-            PreparedStatement ptmt = Conn.getPtmt(con,sql);
-            ptmt.setLong(1,id);
-            ptmt.setDate(2,date);
-            ResultSet rs = ptmt.executeQuery();
-            if(rs.next()){
-                double speed = rs.getDouble("sudu");
-                int num = rs.getInt("typenum");
-                if(speed<Com[0]){
-                    sql = "update allgroupsaiwenchengji set sudu=?,keyspeed=?,keylength=?,groupid=?,typenum=?,name=?,del=?,deletetext=?,sel=?,mistake=?,rightkeyper=? where id=? and saiwendate=?";
-                    ptmt = Conn.getPtmt(con,sql);
-                    ptmt.setDouble(1,Com[0]);
-                    ptmt.setDouble(2,Com[1]);
-                    ptmt.setDouble(3,Com[2]);
-                    ptmt.setLong(4,groupid);
-                    ptmt.setInt(5,num+1);
-                    ptmt.setString(6,username);
-                    ptmt.setInt(7,(int)Com[3]);
-                    ptmt.setInt(8,(int)Com[4]);
-                    ptmt.setInt(9,(int)Com[5]);
-                    ptmt.setInt(10,(int)Com[6]);
-                    ptmt.setDouble(11,Com[7]);
-                    ptmt.setLong(12,id);
-                    ptmt.setDate(13,date);
-                    System.out.println("更改联赛成绩"+id+" "+groupid);
-                    ptmt.execute();
-                }else{
-                    sql = "update allgroupsaiwenchengji set typenum=? where id=? and saiwendate=?";
-                    ptmt = Conn.getPtmt(con,sql);
-                    ptmt.setInt(1,num+1);
-                    ptmt.setLong(2,id);
-                    ptmt.setDate(3,Conn.getdate());
-                    ptmt.execute();
-                }
-            }
-            else{
-                sql = "insert into allgroupsaiwenchengji values(?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
-                ptmt = Conn.getPtmt(con,sql);
-                ptmt.setLong(1,id);
-                ptmt.setLong(2,groupid);
-                ptmt.setDate(3,Conn.getdate());
-                ptmt.setDouble(4,Com[0]);
-                ptmt.setDouble(5,Com[1]);
-                ptmt.setDouble(6,Com[2]);
-                ptmt.setInt(7,1);
-                ptmt.setString(8,username);
-                ptmt.setInt(9,(int)Com[3]);
-                ptmt.setInt(10,(int)Com[4]);
-                ptmt.setInt(11,(int)Com[5]);
-                ptmt.setInt(12,(int)Com[6]);
-                ptmt.setDouble(13,Com[7]);
-                ptmt.setDouble(14,Com[0]);
-                ptmt.execute();
-                System.out.println("添加联赛成绩"+id+" 群"+groupid);
-            }
-            con.close();
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-    }
     public static String addRobotSaiwenMath(Long id,Long groupid,double[] Com,String name,String typeandaid){
         String message = "";
         String []s = typeandaid.split("%");
@@ -452,39 +191,5 @@ public class InConn {
             e.printStackTrace();
         }
         return message;
-    }
-    public static String Deleterobotinfo(long id){
-        String deletechengji  = "delete from allgroupsaiwenchengji where id="+id;
-        String deleteinfo = "delete from robotclient where id="+id;
-        String deletehistory = "delete from robothistory where id="+id;
-        try{
-            Connection connection = Conn.getConnection();
-            Conn.execute(connection,deletechengji);
-            Conn.execute(connection,deleteinfo);
-            Conn.execute(connection,deletehistory);
-            return "删除所有信息包括群跟打记录，历史联赛成绩成功。";
-        }catch (Exception e){
-            e.printStackTrace();
-            return "清零失败";
-        }
-    }
-    public static String  setName(String name,Long id){
-        try{
-            String sql = "update robotclient set name=? where id=?";
-            String sql1 = "select * from robotclient where name=?";
-            Connection con = Conn.getConnection();
-            PreparedStatement ptmt = Conn.getPtmt(con,sql1);
-            ptmt.setString(1,name);
-            ResultSet rs = ptmt.executeQuery();
-            if(rs.next()){con.close();return "已存在相同用户名 Q号为："+rs.getLong("id");}
-            ptmt = Conn.getPtmt(con,sql);
-            ptmt.setString(1,name);
-            ptmt.setLong(2,id);
-            ptmt.execute();
-            con.close();
-            return "更改成功 账号："+id+" 名片："+name;
-        }catch (Exception e){
-            return "更改失败，先打一篇赛文上屏成绩";
-        }
     }
 }
